backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Config
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.0

# Document parsers
pymupdf==1.25.1
python-docx==1.1.2
beautifulsoup4==4.12.3
markdownify==0.14.1

# OCR (OCR.space API client + PDF-to-image renderer)
requests==2.32.3
pdf2image==1.17.0
Pillow==11.1.0

# Email parsing
# (uses stdlib `email` module — no extra dependency needed)

# File type detection
python-magic==0.4.27

# LLM providers
openai==1.58.1
anthropic==0.40.0
google-generativeai==0.8.3
httpx==0.28.1

# Tokenizers
tiktoken==0.8.0

# Utilities
python-ulid==3.0.0
